Make `Store.filter` internal (#1882)
* Make Store.filter internal

We're seeing that `@_spi` can still come up in autocomplete. Now Xcode
surfaces internal stuff too occasionally, but maybe this will suppress
it a bit.

* wip

* fix

(cherry picked from commit b690a617d1366bd36f047e5da5d3185f20daac71)
