+++ Tests/ComposableArchitectureTests/StoreTests.swift
@@ -588,22 +588,5 @@ final class StoreTests: XCTestCase {
 
     XCTAssertEqual(viewStore.state, UUID(uuidString: "deadbeef-dead-beef-dead-beefdeadbeef")!)
   }
-
-  func testFilter() {
-    let store = Store<Int?, Void>(initialState: nil, reducer: EmptyReducer())
-      .filter { state, _ in state != nil }
-
-    let viewStore = ViewStore(store)
-    var count = 0
-    viewStore.publisher
-      .sink { _ in count += 1 }
-      .store(in: &self.cancellables)
-
-    XCTAssertEqual(count, 1)
-    viewStore.send(())
-    XCTAssertEqual(count, 1)
-    viewStore.send(())
-    XCTAssertEqual(count, 1)
-  }
 }
 #endif
